app/requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1 ; python_version < '3.11'
iniconfig==2.0.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
psycopg2-binary==2.9.6
pyjwt==2.6.0 ; python_version >= '3.7'
pytest==7.3.1 ; python_version >= '3.7'
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.3 ; python_version >= '3.5'
tomli==2.0.1 ; python_version < '3.11'
django-money==3.1.0